refactor: replace various any with types and simpler interface naming
