+++ packages/core-transactions/src/handlers/second-signature.ts
@@ -33,7 +33,7 @@ export class SecondSignatureTransactionHandler extends TransactionHandler {
         wallet.secondPublicKey = null;
     }
 
-    public canEnterTransactionPool(data: ITransactionData, guard: TransactionPool.ITransactionGuard): boolean {
+    public canEnterTransactionPool(data: ITransactionData, guard: TransactionPool.IGuard): boolean {
         return !this.typeFromSenderAlreadyInPool(data, guard);
     }
 }
