+++ packages/core-transactions/src/handlers/transfer.ts
@@ -34,7 +34,7 @@ export class TransferTransactionHandler extends TransactionHandler {
         return;
     }
 
-    public canEnterTransactionPool(data: ITransactionData, guard: TransactionPool.ITransactionGuard): boolean {
+    public canEnterTransactionPool(data: ITransactionData, guard: TransactionPool.IGuard): boolean {
         if (!isRecipientOnActiveNetwork(data)) {
             guard.pushError(
                 data,
